update
1. Task #13
2. Task #31

+++ task31/task31.h
@@ -6,7 +6,6 @@
 void start(void);//начало
 void task(void);//задание
 void Beep(void);
-u_int mas_to_int(short *mas, size_t size);
-u_int power(u_int num, short ex);
-
+unsigned long int mas_to_int(short *mas, size_t size);
+unsigned long int power(u_int num, short ex);
 #endif
